Split test into two cases.
One is to test our core implementation. Most of the current tests are this.
They are moved in test/core directory.

Another case is to test the library we provides. Theya re in test/lib
directory. But most of our library just exports the core implementation. So I
don't add test cases for them. We will only test the libraries that depend on
external libraries(e.g. future and uuid) or not in the core
implementation(e.g. db).