Add `contains` operator to Query class.
A `contains` method has the same behavior of `$all` operator
that selects the documents where the value of a field
is an array that contains all the specified elements.
As there already exists `all` operator and it's totally
different, I made the `contains` method that possibly has
more clear meaning.

http://docs.mongodb.org/manual/reference/operator/query/all