Call plugin's startup method after http server starting.
If it is exist public method named 'startup' in plugin's main script,
call it after http server starting.

When the server startup, it is need a way to perform tasks that are required for plugin additionaly.
Currently, it is unable to know when the server start listening after plugin's initialization is complete.
Therefore, the server startup after its initialization is complete, call plugin's startup method if it is exist.